+++ src/Pages/Dashboard/AdminDashboard/AdminDashboardHome.js
@@ -7,59 +7,12 @@ import { AuthContext } from "../../../contexts/AuthProvider";
 import axios from "axios";
 
 import { CircularProgress } from "@mui/material";
-import { useLocation, useNavigate } from "react-router-dom";
-import { jwtVerify } from "jose";
-import Loading from "../../Shared/Loading/Loading";
-function useQuery() {
-  return new URLSearchParams(useLocation().search);
-}
 
 const AdminDashboardHome = () => {
-  const { userInfo, signIn } = useContext(AuthContext);
+  const { userInfo } = useContext(AuthContext);
   const [organization, setOrganization] = useState();
   const [isLoading, setIsLoading] = useState(true);
   const [itemDetails, setItemDetails] = useState();
-  const query = useQuery();
-  const token = query.get("token");
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const getTokenData = async () => {
-      const newSecret = process.env.REACT_APP_SECRET_LOGIN_KEY;
-      const secret = new TextEncoder().encode(newSecret);
-      try {
-        if (secret && token) {
-          const { payload } = await jwtVerify(token, secret);
-          if (!payload.isProviderLogin) {
-            if (!userInfo.email) {
-              await signIn(payload.email, payload.password).then(() => {
-                saveUser(payload.email);
-              });
-            }
-          }
-        }
-      } catch (error) {
-        console.log({ error });
-        throw new Error(error);
-      }
-    };
-
-    token && getTokenData();
-  }, [token]);
-
-  const saveUser = async (email) => {
-    try {
-      fetch(
-        `${process.env.REACT_APP_SERVERLESS_API}/api/v1/users?email=${email}`
-      )
-        .then((res) => res.json())
-        .then((data) => {
-          localStorage.setItem("role", data?.role);
-          if (data?.role === "admin") navigate("/adminDashboardHome");
-          else navigate("/dashboard");
-        });
-    } catch (error) {}
-  };
 
   useEffect(() => {
     axios
